+++ packages/terra-search-field/src/SearchField.jsx
@@ -9,7 +9,16 @@ import styles from './SearchField.scss';
 
 const cx = classNames.bind(styles);
 
+const KEYCODES = {
+  ENTER: 13,
+};
+
 const propTypes = {
+  /**
+   * When true, will disable the auto-search
+   */
+  disableAutoSearch: PropTypes.bool,
+
   /**
    * Placeholder text to show while the search field is empty.
    */
@@ -37,6 +46,7 @@ const propTypes = {
 };
 
 const defaultProps = {
+  disableAutoSearch: false,
   placeholder: '',
   minimumSearchTextLength: 2,
   searchDelay: 250,
@@ -49,6 +59,7 @@ class SearchField extends React.Component {
 
     this.handleTextChange = this.handleTextChange.bind(this);
     this.handleSearch = this.handleSearch.bind(this);
+    this.handleKeyDown = this.handleKeyDown.bind(this);
 
     this.searchTimeout = null;
     this.state = {
@@ -63,8 +74,14 @@ class SearchField extends React.Component {
   handleTextChange(event) {
     this.setState({ searchText: event.target.value });
 
-    if (!this.searchTimeout) {
+    if (!this.searchTimeout && !this.props.disableAutoSearch) {
       this.searchTimeout = setTimeout(this.handleSearch, this.props.searchDelay);
+    }
+  }
+
+  handleKeyDown(event) {
+    if (event.nativeEvent.keyCode === KEYCODES.ENTER) {
+      this.handleSearch();
     }
   }
 
@@ -86,7 +103,7 @@ class SearchField extends React.Component {
   }
 
   render() {
-    const { placeholder, searchDelay, minimumSearchTextLength, onSearch, onInvalidSearch, ...customProps } = this.props;
+    const { placeholder, searchDelay, minimumSearchTextLength, onSearch, onInvalidSearch, disableAutoSearch, ...customProps } = this.props;
     const searchFieldClassNames = cx([
       'searchfield',
       customProps.className,
@@ -100,6 +117,7 @@ class SearchField extends React.Component {
           placeholder={placeholder}
           value={this.state.searchText}
           onChange={this.handleTextChange}
+          onKeyDown={this.handleKeyDown}
         />
         <Button
           className={cx('button')}