+++ packages/terra-search-field/tests/nightwatch/SearchFieldTestRoutes.jsx
@@ -8,6 +8,8 @@ import PlaceholderSearchField from './PlaceholderSearchField';
 import CallbackSearchField from './CallbackSearchField';
 import DelayedSearchField from './DelayedSearchField';
 import MinimumLengthSearchField from './MinimumLengthSearchField';
+import AutoSearchDisabledSearchField from './AutoSearchDisabledSearchField';
+import SearchFieldEnter from './SearchFieldEnter';
 
 const routes = (
   <div>
@@ -17,6 +19,8 @@ const routes = (
     <Route path="/tests/search-field-tests/callback" component={CallbackSearchField} />
     <Route path="/tests/search-field-tests/delayed" component={DelayedSearchField} />
     <Route path="/tests/search-field-tests/minimum-length" component={MinimumLengthSearchField} />
+    <Route path="/tests/search-field-tests/disable-auto-search" component={AutoSearchDisabledSearchField} />
+    <Route path="/tests/search-field-tests/enter" component={SearchFieldEnter} />
   </div>
 );